requirements.txt
tqdm==2.2.3
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
Pillow==10.1.0
matplotlib==3.8.2